+++ packages/desktop-client/src/components/sidebar/Accounts.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { moveAccount } from 'loot-core/client/accounts/accountSlice';
+import { moveAccount } from 'loot-core/client/accounts/accountsSlice';
 import * as queries from 'loot-core/client/queries';
 import { type AccountEntity } from 'loot-core/types/models';
